pkcs12 failed. rolled back changes

+++ create-ssl-cert.sh
@@ -71,22 +71,22 @@ echo "Two letter Country Codes: https://www.digicert.com/ssl-certificate-country
 echo
 read -n1 -r -p "Press any key to continue..."
 
-#openssl req -new -newkey rsa:2048 -nodes -sha256 -x509 -days 365 -keyout ~/server.key -out ~/server.crt
 
-openssl req -sha256 -x509 -newkey rsa:2048 -keyout ~/key.pem -out ~/server.pem -days 365
-#if [ ! -e ~/server.key ] || [ ! -e ~/server.crt ]; then
-  #echo "Error creation of SSL certificate has failed.  Aborting"
-  #exit 2
-#fi
+openssl req -new -newkey rsa:2048 -nodes -sha256 -x509 -days 365 -keyout ~/server.key -out ~/server.crt
+if [ ! -e ~/server.key ] || [ ! -e ~/server.crt ]; then
+  echo "Error creation of SSL certificate has failed.  Aborting"
+  exit 2
+fi
 
-#echo "Merging Crt file and Key file to form Pem"
-#cat ~/server.key ~/server.crt > ~/server.pem
-#echo
-
+echo "Merging Crt file and Key file to form Pem"
+cat ~/server.key ~/server.crt > ~/server.pem
 echo
-echo "Generating pkcs12 certificate"
-echo "The pass phrase is what you just typed in earlier"
-openssl pkcs12 -export -in ~/server.pem -inkey ~/key.pem -name "$HostName" -out "$HostName-cert.p12"
+
+#pkcs12 method doesn't work in Lighttpd 
+#openssl req -sha256 -x509 -newkey rsa:2048 -keyout ~/key.pem -out ~/server.pem -days 365
+#echo "Generating pkcs12 certificate"
+#echo "The pass phrase is what you just typed in earlier"
+#openssl pkcs12 -export -in ~/server.pem -inkey ~/key.pem -name "$HostName" -out "$HostName-cert.p12"
 
 echo "Copying Certificate to /etc/lighttpd/"
 sudo cp ~/server.pem /etc/lighttpd/server.pem
